@@ -56,44 +60,98 @@
     },
     "homepage": "https://github.com/ressurectit/ng-rest",
     "peerDependencies": {
-        "@angular/core": ">=10.0.0",
-        "@angular/common": ">=10.0.0",
-        "rxjs": ">=6.5.0",
-        "@jscrpt/common": ">=1.2.0",
-        "@anglr/common": ">=9.0.0",
+        "@angular/core": ">=13.1.0",
+        "@angular/common": ">=13.1.0",
+        "rxjs": ">=6.6.7",
+        "@jscrpt/common": ">=2.2.0",
+        "@anglr/common": ">=10.0.0",
         "crypto-js": "^4.0.0",
         "moment": "^2.29.1",
-        "date-fns": "^2.19.0",
+        "date-fns": "^2.27.0",
         "@stomp/stompjs": "^5.4.2",
         "sockjs-client": "^1.3.0",
-        "avsc": "^5.4.21"
+        "avsc": "^5.4.21",
+        "tslib": "^2.3.1"
     },
     "devDependencies": {
-        "@angular/core": "^10.0.0",
-        "@angular/common": "^10.0.0",
-        "@angular/platform-browser": "^10.0.0",
-        "rxjs": "^6.5.0",
-        "@jscrpt/common": "^1.2.0",
-        "@anglr/common": "^9.0.0",
+        "@angular/core": "^13.1.0",
+        "@angular/common": "^13.1.0",
+        "@angular/platform-browser": "^13.1.0",
+        "rxjs": "^6.6.7",
+        "@jscrpt/common": "^2.2.0",
+        "@anglr/common": "^10.0.0",
         "moment": "^2.29.1",
-        "date-fns": "^2.19.0",
+        "date-fns": "^2.27.0",
         "avsc": "^5.4.21",
-        "@anglr/types": "1.1.18",
-        "@types/node": "11.11.1",
+        "@types/node": "^14.14.31",
         "@stomp/stompjs": "^5.4.2",
-        "@types/crypto-js": "3.1.47",
+        "@types/crypto-js": "4.0.2",
         "@types/sockjs-client": "^1.1.1",
-        "@angular/compiler-cli": "^10.0.0",
-        "@angular/compiler": "^10.0.0",
-        "@angular/language-service": "^10.0.0",
-        "@microsoft/api-extractor": "7.9.1",
-        "@microsoft/api-documenter": "7.8.21",
-        "tsickle": "0.38.0",
+        "@angular/compiler-cli": "^13.1.0",
+        "@angular/compiler": "^13.1.0",
+        "@angular/language-service": "^13.1.0",
+        "@microsoft/api-extractor": "7.19.2",
+        "@microsoft/api-documenter": "7.13.77",
+        "tslib": "2.3.1",
+        "eslint": "7.32.0",
+        "@typescript-eslint/eslint-plugin": "4.33.0",
+        "@typescript-eslint/parser": "4.33.0",
+        "eslint-plugin-ressurectit": "0.1.0",
         "rimraf": "3.0.2",
-        "copyfiles": "2.3.0",
-        "replace-in-file": "6.1.0",
-        "path-exists-cli": "1.0.0",
-        "typescript": "3.9.7",
-        "npm-git-version": "1.1.1"
+        "copyfiles": "2.4.1",
+        "replace-in-file": "6.3.2",
+        "path-exists-cli": "2.0.0",
+        "typescript": "4.5.4",
+        "npm-git-version": "2.0.0"
+    },
+    "esm2020": "./es2020/src/index.js",
+    "es2020": "./es2020/src/index.js",
+    "es2015": "./es2015/src/index.js",
+    "module": "./es2015/src/index.js",
+    "typings": "./rest.d.ts",
+    "exports": {
+        "./package.json": {
+            "default": "./package.json"
+        },
+        ".": {
+            "types": "./rest.d.ts",
+            "esm2020": "./es2020/src/index.js",
+            "es2020": "./es2020/src/index.js",
+            "es2015": "./es2015/src/index.js",
+            "node": "./es2015/src/index.js",
+            "default": "./es2020/src/index.js"
+        },
+        "./date-fns": {
+            "types": "./date-fns/rest-date-fns.d.ts",
+            "esm2020": "./es2020/date-fns/src/index.js",
+            "es2020": "./es2020/date-fns/src/index.js",
+            "es2015": "./es2015/date-fns/src/index.js",
+            "node": "./es2015/date-fns/src/index.js",
+            "default": "./es2020/date-fns/src/index.js"
+        },
+        "./moment": {
+            "types": "./moment/rest-moment.d.ts",
+            "esm2020": "./es2020/moment/src/index.js",
+            "es2020": "./es2020/moment/src/index.js",
+            "es2015": "./es2015/moment/src/index.js",
+            "node": "./es2015/moment/src/index.js",
+            "default": "./es2020/moment/src/index.js"
+        },
+        "./avsc": {
+            "types": "./avsc/rest-avsc.d.ts",
+            "esm2020": "./es2020/avsc/src/index.js",
+            "es2020": "./es2020/avsc/src/index.js",
+            "es2015": "./es2015/avsc/src/index.js",
+            "node": "./es2015/avsc/src/index.js",
+            "default": "./es2020/avsc/src/index.js"
+        },
+        "./stompjs": {
+            "types": "./stompjs/rest-stompjs.d.ts",
+            "esm2020": "./es2020/stompjs/src/index.js",
+            "es2020": "./es2020/stompjs/src/index.js",
+            "es2015": "./es2015/stompjs/src/index.js",
+            "node": "./es2015/stompjs/src/index.js",
+            "default": "./es2020/stompjs/src/index.js"
+        }
     }
 }
