@@ -1,15 +1,15 @@
 import {Injectable, Inject, Optional, ClassProvider} from '@angular/core';
 import {HttpInterceptor, HTTP_INTERCEPTORS, HttpEvent, HttpHandler, HttpRequest, HttpEventType} from '@angular/common/http';
-import {IgnoredInterceptorsService, IgnoredInterceptorId, AdditionalInfo} from '@anglr/common';
+import {IGNORED_INTERCEPTORS} from '@anglr/common';
 import {HTTP_HEADER_CONTENT_TYPE} from '@anglr/rest';
+import {StringDictionary} from '@jscrpt/common';
 import {Observable} from 'rxjs';
 import {map} from 'rxjs/operators';
 import {Schema, Type} from 'avsc';
 
 import {AvroAdapterInterceptorOptions} from './avroAdapter.options';
-import {AVRO_ADAPTER_SCHEMA_PROVIDER} from '../../misc/tokens';
+import {AVRO_ADAPTER_SCHEMA_PROVIDER, AVRO_REQUEST_DATA, AVRO_RESPONSE_DATA} from '../../misc/tokens';
 import {AvroAdapterSchemaProvider} from '../../services/avroAdapterSchemaProvider/avroAdapterSchemaProvider.interface';
-import {AvroResponseType, AvroRequestType} from './avroAdapter.interface';
 
 /**
  * Interceptor that will enable usage of AVRO for request and response data streams (binary format)
@@ -19,7 +19,6 @@ export class AvroAdapterInterceptor implements HttpInterceptor
 {
     //######################### constructor #########################
     constructor(@Optional() private _options: AvroAdapterInterceptorOptions,
-                @Optional() private _ignoredInterceptorsService: IgnoredInterceptorsService,
                 @Optional() @Inject(AVRO_ADAPTER_SCHEMA_PROVIDER) private _schemaProvider: AvroAdapterSchemaProvider)
     {
         if(!_options)
@@ -35,11 +34,17 @@ export class AvroAdapterInterceptor implements HttpInterceptor
      * @param req - Request to be intercepted
      * @param next - Next middleware that can be called for next processing
      */
-    public intercept(req: HttpRequest<any> & AdditionalInfo<IgnoredInterceptorId & AvroRequestType & AvroResponseType>, next: HttpHandler): Observable<HttpEvent<any>>
+    public intercept(req: HttpRequest<any>, next: HttpHandler): Observable<HttpEvent<any>>
     {
-        let avroReq = req.additionalInfo?.avroRequest;
-        let avroRes = req.additionalInfo?.avroResponse;
-        let schemaObj = this._schemaProvider.schema;
+        //interceptor is ignored
+        if(req.context.get(IGNORED_INTERCEPTORS).some(itm => itm == AvroAdapterInterceptor))
+        {
+            return next.handle(req);
+        }
+
+        const avroReq = req.context.get(AVRO_REQUEST_DATA);
+        const avroRes = req.context.get(AVRO_RESPONSE_DATA);
+        const schemaObj = this._schemaProvider.schema;
 
         if(this._options.disabled ||
            (!avroReq && !avroRes))
@@ -55,12 +60,12 @@ export class AvroAdapterInterceptor implements HttpInterceptor
             //body present and schema for specified type exists and content type provided
             if(req.body && schemaObj[avroReq.namespace] && (schema = schemaObj[avroReq.namespace][avroReq.name]) && this._options.customAcceptContentTypeHeader)
             {
-                let type = Type.forSchema(schema);
-                let additionalHeaders = {};
+                const type = Type.forSchema(schema);
+                const additionalHeaders: StringDictionary = {};
 
                 additionalHeaders[HTTP_HEADER_CONTENT_TYPE] = this._options.customAcceptContentTypeHeader;
 
-                if(this._options.typeHeaderName)
+                if(this._options.typeHeaderName && type.name)
                 {
                     additionalHeaders[this._options.typeHeaderName] = type.name;
                 }
@@ -94,12 +99,6 @@ export class AvroAdapterInterceptor implements HttpInterceptor
         return next.handle(req)
             .pipe(map(result =>
                       {
-                          //is ignored
-                          if (this._ignoredInterceptorsService && this._ignoredInterceptorsService.isIgnored(AvroAdapterInterceptor, req.additionalInfo))
-                          {
-                              return result;
-                          }
-                          
                           //process only response
                           if(result.type != HttpEventType.Response)
                           {
@@ -112,11 +111,11 @@ export class AvroAdapterInterceptor implements HttpInterceptor
                           if(avroRes && result.headers.get(HTTP_HEADER_CONTENT_TYPE) == this._options.customAcceptContentTypeHeader &&
                              schemaObj[avroRes.namespace] && (schema = schemaObj[avroRes.namespace][avroRes.name]))
                           {
-                              let type = Type.forSchema(schema);
+                              const type = Type.forSchema(schema);
 
                               return result.clone<any>(
                               {
-                                  body: type.fromBuffer(new Buffer(new Uint8Array(result.body)))
+                                  body: type.fromBuffer(Buffer.from(new Uint8Array(result.body)))
                               });
                           }
 
