@@ -1,14 +1,14 @@
-import {Injector} from "@angular/core";
-import {generateId, extend, isPresent} from "@jscrpt/common";
-import {Logger} from "@anglr/common";
-import {Client} from "@stomp/stompjs";
-import {ReplaySubject, MonoTypeOperatorFunction, Observable} from "rxjs";
-import {tap} from "rxjs/operators";
+import {Injector} from '@angular/core';
+import {Logger} from '@anglr/common';
+import {generateId, extend, isPresent} from '@jscrpt/common';
+import {Client} from '@stomp/stompjs';
+import {tap} from 'rxjs/operators';
+import {ReplaySubject, MonoTypeOperatorFunction, Observable} from 'rxjs';
 
-import {WebSocketClientResponse, StatusQueueResponse, WebSocketHandleResultMiddleware, WebSocketHandleStatusSubscribeMiddleware} from "./webSocketClient.interface";
-import {WebSocketClientResponseOptions, SubscriptionMetadata, WebSocketClientOptions} from "./webSocketClient.interface.internal";
-import {RequestType, QueueCorrelationPosition} from "./webSocketClient.types";
-import {getCache, storeToCache} from "./webSocketClinet.cache";
+import {WebSocketClientResponse, StatusQueueResponse, WebSocketHandleResultMiddleware, WebSocketHandleStatusSubscribeMiddleware, PublishQueueOptions} from './webSocketClient.interface';
+import {WebSocketClientResponseOptions, SubscriptionMetadata, WebSocketClientOptions} from './webSocketClient.interface.internal';
+import {RequestType, QueueCorrelationPosition} from './webSocketClient.types';
+import {getCache, storeToCache} from './webSocketClinet.cache';
 
 const DEFAULT_STATUS_MAPPING = (value: any) => value;
 
@@ -53,7 +53,7 @@ export class WebSocketClientResponseContext implements WebSocketClientResponse<a
     /**
      * Key used to identifying cache
      */
-    private _cacheKey: string;
+    private _cacheKey!: string;
 
     //######################### public properties - implementation of WebSocketClientResponse<any> #########################
 
@@ -85,7 +85,7 @@ export class WebSocketClientResponseContext implements WebSocketClientResponse<a
     /**
      * Explicitly call publish to queue, can be called only once, if called multiple times, rest of calls do nothing
      */
-    public publish()
+    public publish(): void
     {
         if(this._published)
         {
@@ -93,27 +93,27 @@ export class WebSocketClientResponseContext implements WebSocketClientResponse<a
         }
 
         this._published = true;
-        let options = this._getOptions(this._options.options);
+        const options = this._getOptions(this._options.options);
         let body = this._options.body;
 
-        if(options.cacheResponse)
+        if(options?.cacheResponse)
         {
-            this._cacheKey = this._buildCacheKey(options.publishQueuePrefix, this._options.publish, options.cacheResponse(body));
+            this._cacheKey = this._buildCacheKey(options.publishQueuePrefix!, this._options.publish, options.cacheResponse(body));
         }
 
-        if(isPresent(body) && options.type == RequestType.Json)
+        if(isPresent(body) && options?.type == RequestType.Json)
         {
             if(isPresent(options.correlationBodyProperty))
             {
-                body[options.correlationBodyProperty] = this._correlationId;
+                (body as any)[options.correlationBodyProperty] = this._correlationId;
             }
 
             body = JSON.stringify(body);
         }
 
-        if(options.cacheResponse)
+        if(options?.cacheResponse)
         {
-            let cache = getCache(this._cacheKey);
+            const cache = getCache(this._cacheKey);
 
             if(cache)
             {
@@ -125,11 +125,11 @@ export class WebSocketClientResponseContext implements WebSocketClientResponse<a
 
         this._logger.debug(`WebSocket: sending request to '${this._options.publish}' with '${body}'`);
 
-        jsDevMode && WS_DEBUG_OPTIONS.request && console.log(`DEBUG REQUEST ${this._generateQueuePublishUrl(options.publishQueuePrefix, this._options.publish, options)} => ${body}`);
+        jsDevMode && WS_DEBUG_OPTIONS.request && console.log(`DEBUG REQUEST ${this._generateQueuePublishUrl(options?.publishQueuePrefix, this._options.publish, options)} => ${body}`);
 
         this._wsClient.publish(
         {
-            destination: this._generateQueuePublishUrl(options.publishQueuePrefix, this._options.publish, options),
+            destination: this._generateQueuePublishUrl(options?.publishQueuePrefix, this._options.publish, options),
             body: body as string
         });
     }
@@ -145,12 +145,12 @@ export class WebSocketClientResponseContext implements WebSocketClientResponse<a
 
             Object.keys(this._outputMetadata).forEach(name =>
             {
-                let metadata = this._outputMetadata[name];
+                const metadata = this._outputMetadata[name];
 
                 if(metadata.subscription)
                 {
                     metadata.subscription.unsubscribe();
-                    metadata.subscription = null;
+                    metadata.subscription = null!;
                 }
 
                 if(metadata.subject)
@@ -170,7 +170,7 @@ export class WebSocketClientResponseContext implements WebSocketClientResponse<a
     {
         Object.keys(this._options.subscribe).forEach(name =>
         {
-            let subject = new ReplaySubject<any>();
+            const subject = new ReplaySubject<any>();
 
             this.output[name] = subject
                 .asObservable()
@@ -182,34 +182,34 @@ export class WebSocketClientResponseContext implements WebSocketClientResponse<a
             };
         });
 
-        let publishOptions = this._getOptions(this._options.options);
+        const publishOptions = this._getOptions(this._options.options);
 
         await this._active;
 
         Object.keys(this._options.subscribe).forEach(name =>
         {
-            let metadata = this._options.subscribe[name];
-            let outputMetadata = this._outputMetadata[name];
-            let options = this._getOptions(metadata.options);
+            const metadata = this._options.subscribe[name];
+            const outputMetadata = this._outputMetadata[name];
+            const options = this._getOptions(metadata.options);
 
-            outputMetadata.subscription = this._wsClient.subscribe(this._generateQueuePublishUrl(options.subscribeQueuePrefix, metadata.queueName, options), data =>
+            outputMetadata.subscription = this._wsClient.subscribe(this._generateQueuePublishUrl(options?.subscribeQueuePrefix, metadata.queueName, options), data =>
             {
                 this._logger.debug(`WebSocket: subscription "${this._sessionId}", queue "${name}", data "${data.body}"`);
-                jsDevMode && WS_DEBUG_OPTIONS.responseSubscribe && console.log(`DEBUG SUBSCRIBE RAW ${this._generateQueuePublishUrl(publishOptions.publishQueuePrefix, this._options.publish, publishOptions)} => ${name}`, data.body);
+                jsDevMode && WS_DEBUG_OPTIONS.responseSubscribe && console.log(`DEBUG SUBSCRIBE RAW ${this._generateQueuePublishUrl(publishOptions?.publishQueuePrefix, this._options.publish, publishOptions)} => ${name}`, data.body);
 
-                if(publishOptions.cacheResponse)
+                if(publishOptions?.cacheResponse)
                 {
                     storeToCache(this._cacheKey, name, data);
                 }
 
                 //handle status queue
-                if(options.statusQueue)
+                if(options?.statusQueue)
                 {
                     let statusData: any = data.body;
 
                     try
                     {
-                        statusData = JSON.parse(data.body)
+                        statusData = JSON.parse(data.body);
                     }
                     catch(e)
                     {
@@ -217,25 +217,25 @@ export class WebSocketClientResponseContext implements WebSocketClientResponse<a
                         statusData = {};
                     }
 
-                    let status: StatusQueueResponse = options.statusQueueMapping ? options.statusQueueMapping(statusData) : DEFAULT_STATUS_MAPPING(statusData);
+                    const status: StatusQueueResponse = options.statusQueueMapping ? options.statusQueueMapping(statusData) : DEFAULT_STATUS_MAPPING(statusData);
                     // status.original = data;
 
                     this._handleStatusMiddlewares.forEach(middleware =>
                     {
-                        middleware(status, metadata, options, publishOptions, this._outputMetadata[status.queue], this._injector, this._correlationId, name);
+                        middleware(status, metadata, options, publishOptions!, this._outputMetadata[status.queue!], this._injector, this._correlationId, name);
                     });
                 }
 
                 outputMetadata.subject.next(data);
 
-                if(options.singleResponse)
+                if(options?.singleResponse)
                 {
                     outputMetadata.subject.complete();
                 }
             });
         });
 
-        if(!this._options.options.delayed)
+        if(!this._options.options?.delayed)
         {
             this.publish();
         }
@@ -257,14 +257,14 @@ export class WebSocketClientResponseContext implements WebSocketClientResponse<a
      */
     private _generateQueuePublishUrl(prefix?: string, name?: string, options?: WebSocketClientOptions)
     {
-        if(options.queueCorrelation && options.queueCorrelation.position == QueueCorrelationPosition.Prefix)
+        if(options?.queueCorrelation && options.queueCorrelation.position == QueueCorrelationPosition.Prefix)
         {
             prefix = `${prefix}/${this._correlationId}`;
         }
 
-        let url = `${prefix}/${name}${this._generateSuffix(options)}`;
+        let url = `${prefix}/${name}${this._generateSuffix(options!)}`;
 
-        if(options.queueCorrelation && options.queueCorrelation.position == QueueCorrelationPosition.Replace)
+        if(options?.queueCorrelation && options.queueCorrelation.position == QueueCorrelationPosition.Replace)
         {
             url = url.replace(`{${options.queueCorrelation.replacementKey}}`, this._correlationId);
         }
@@ -278,7 +278,7 @@ export class WebSocketClientResponseContext implements WebSocketClientResponse<a
      */
     private _generateSuffix(options: WebSocketClientOptions): string
     {
-        let suffix = "";
+        let suffix = '';
 
         if(options.sessionIdSuffix)
         {
@@ -308,9 +308,9 @@ export class WebSocketClientResponseContext implements WebSocketClientResponse<a
      */
     private _handleResult: (name: string) => MonoTypeOperatorFunction<any> = name =>
     {
-        let metadata = this._options.subscribe[name];
-        let options = this._getOptions(metadata.options);
-        let publishOptions = jsDevMode && this._getOptions(this._options.options);
+        const metadata = this._options.subscribe[name];
+        const options = this._getOptions(metadata.options);
+        const publishOptions = (jsDevMode && this._getOptions(this._options.options)) as PublishQueueOptions | undefined;
 
         return (source: Observable<any>) =>
         {
@@ -318,20 +318,20 @@ export class WebSocketClientResponseContext implements WebSocketClientResponse<a
             {
                 source = source.pipe(tap(data =>
                 {
-                    console.log(`DEBUG RAW ${this._generateQueuePublishUrl(publishOptions.publishQueuePrefix, this._options.publish, publishOptions)} => ${name}`, data.body);
+                    console.log(`DEBUG RAW ${this._generateQueuePublishUrl(publishOptions?.publishQueuePrefix, this._options.publish, publishOptions)} => ${name}`, data.body);
                 }));
             }
 
             this._handleResultMiddlewares.forEach(middleware =>
             {
-                source = middleware(source, metadata, options, publishOptions, this._injector, this._correlationId, name);
+                source = middleware(source, metadata, options!, publishOptions!, this._injector, this._correlationId, name);
             });
 
             if(jsDevMode && WS_DEBUG_OPTIONS.response)
             {
                 source = source.pipe(tap(data =>
                 {
-                    console.log(`DEBUG ${this._generateQueuePublishUrl(publishOptions.publishQueuePrefix, this._options.publish, publishOptions)} => ${name}`, data);
+                    console.log(`DEBUG ${this._generateQueuePublishUrl(publishOptions?.publishQueuePrefix, this._options.publish, publishOptions)} => ${name}`, data);
                 }));
             }
 
