@@ -1,4 +1,4 @@
-import {extend} from '@jscrpt/common';
+import {extend, StringDictionary} from '@jscrpt/common';
 
 import {RESTClient} from '../common';
 import {RestHttpHeaders, RestMethodMiddlewares} from '../rest.interface';
@@ -8,13 +8,13 @@ import {HeadersMiddleware} from '../middlewares';
  * Set custom headers for a REST method
  * @param headersDef - custom headers in a key-value pair
  */
-export function Headers(headersDef: {[key: string]: string})
+export function Headers(headersDef: StringDictionary)
 {
     return function(_target: RESTClient, _propertyKey: string, descriptor: RestHttpHeaders &
-                                                                           RestMethodMiddlewares)
+                                                                           RestMethodMiddlewares): TypedPropertyDescriptor<any>
     {
         descriptor.headers = extend(descriptor.headers ?? {}, headersDef);
-        descriptor.middlewareTypes.push(HeadersMiddleware);
+        descriptor.middlewareTypes?.push(HeadersMiddleware);
 
         return descriptor;
     };
