@@ -9,9 +9,9 @@ function paramBuilder(paramName: keyof ParametersMetadata, middleware: Type<Rest
 {
     return function(key: string)
     {
-        return function(target: RESTClient & RestParameters, propertyKey: string, parameterIndex: number)
+        return function(target: RESTClient & RestParameters, propertyKey: string, parameterIndex: number): void
         {
-            let paramObj: KeyIndex =
+            const paramObj: KeyIndex =
             {
                 key: key,
                 parameterIndex: parameterIndex
@@ -27,13 +27,13 @@ function paramBuilder(paramName: keyof ParametersMetadata, middleware: Type<Rest
             target.parameters[propertyKey][paramName] = target.parameters[propertyKey][paramName] ?? [];
 
             //adds params
-            target.parameters[propertyKey][paramName].push(paramObj);
+            target.parameters[propertyKey][paramName]?.push(paramObj);
 
             //sets middleware
             if(isPresent(middleware))
             {
                 target.parameters[propertyKey].middlewareTypes = target.parameters[propertyKey].middlewareTypes ?? [];
-                target.parameters[propertyKey].middlewareTypes.push(middleware);
+                target.parameters[propertyKey].middlewareTypes?.push(middleware);
             }
         };
     };
@@ -43,27 +43,27 @@ function paramBuilder(paramName: keyof ParametersMetadata, middleware: Type<Rest
  * Path variable of a method's url, type: string
  * @param key - path key to bind value
  */
-export var Path = paramBuilder("path", PathParameterMiddleware);
+export const Path = paramBuilder('path', PathParameterMiddleware);
 
 /**
  * Query value of a method's url, type: string
  * @param key - query key to bind value
  */
-export var Query = paramBuilder("query", QueryParameterMiddleware);
+export const Query = paramBuilder('query', QueryParameterMiddleware);
 
 /**
  * Query object serialized with dot notation separating hierarchies
  */
-export var QueryObject = paramBuilder("queryObject", QueryObjectParameterMiddleware)("queryObject");
+export const QueryObject = paramBuilder('queryObject', QueryObjectParameterMiddleware)('queryObject');
 
 /**
  * Body of a REST method, json stringify applied
  * Only one body per method!
  */
-export var Body = paramBuilder("body", BodyParameterMiddleware)("body");
+export const Body = paramBuilder('body', BodyParameterMiddleware)('body');
 
 /**
  * Custom header of a REST method, type: string
  * @param key - header key to bind value
  */
-export var Header = paramBuilder("header", HeaderParameterMiddleware);
+export const Header = paramBuilder('header', HeaderParameterMiddleware);
