@@ -1,7 +1,7 @@
 import {Type} from '@angular/core';
 
 import {RestMiddleware} from './rest.interface';
-import {ReportProgressMiddleware, ResponseTypeMiddleware, IgnoredInterceptorsMiddleware, ResponseTransformMiddleware, AdditionalDataMiddleware, ProducesMiddleware, BodyParameterMiddleware, PathParameterMiddleware, QueryObjectParameterMiddleware, QueryParameterMiddleware, HeaderParameterMiddleware, HeadersMiddleware, CacheMiddleware, LoggerMiddleware} from './middlewares';
+import {ReportProgressMiddleware, ResponseTypeMiddleware, IgnoredInterceptorsMiddleware, ResponseTransformMiddleware, ProducesMiddleware, BodyParameterMiddleware, PathParameterMiddleware, QueryObjectParameterMiddleware, QueryParameterMiddleware, HeaderParameterMiddleware, HeadersMiddleware, CacheMiddleware, LoggerMiddleware, ProgressIndicatorGroupMiddleware} from './middlewares';
 
 /**
  * Definition of basic default array of rest middlewares order
@@ -16,8 +16,8 @@ export const BASIC_DEFAULT_REST_MIDDLEWARES_ORDER: Type<RestMiddleware>[] =
     HeaderParameterMiddleware,
     ProducesMiddleware,
     LoggerMiddleware,
-    AdditionalDataMiddleware,
     IgnoredInterceptorsMiddleware,
+    ProgressIndicatorGroupMiddleware,
     ResponseTransformMiddleware,
     ResponseTypeMiddleware,
     CacheMiddleware,
