@@ -2,7 +2,6 @@ export * from './reportProgress.middleware';
 export * from './responseType.middleware';
 export * from './responseTransform.middleware';
 export * from './ignoredInterceptors.middleware';
-export * from './additionalData.middleware';
 export * from './produces.middleware';
 export * from './bodyParameter.middleware';
 export * from './pathParameter.middleware';
@@ -12,3 +11,4 @@ export * from './headerParameter.middleware';
 export * from './headers.middleware';
 export * from './cache.middleware';
 export * from './logger.middleware';
+export * from './progressIndicatorGroup.middleware';
