@@ -1,13 +1,14 @@
 import {HttpRequest} from '@angular/common/http';
-import {AdditionalInfo, Logger, LOGGER} from '@anglr/common';
+import {Logger, LOGGER} from '@anglr/common';
 import {Observable} from 'rxjs';
 import {tap} from 'rxjs/operators';
 
-import {RestMiddleware, ɵRESTClient} from '../rest.interface';
+import {RestMiddleware} from '../rest.interface';
+import type {RESTClient} from '../common';
 
 interface ɵLogger
 {
-    ɵLogger?: Logger;
+    ɵLogger: Logger|null;
 }
 
 /**
@@ -28,23 +29,24 @@ export class LoggerMiddleware implements RestMiddleware
      * @param request - Http request that you can modify
      * @param next - Used for calling next middleware with modified request
      */
-    public run(this: ɵRESTClient & ɵLogger,
+    public run(this: RESTClient,
                _id: string,
-               _target: any,
+               _target: unknown,
                methodName: string,
-               _descriptor: any,
-               args: any[],
-               request: HttpRequest<any> & AdditionalInfo,
-               next: (request: HttpRequest<any>) => Observable<any>): Observable<any>
+               _descriptor: unknown,
+               args: unknown[],
+               request: HttpRequest<unknown>,
+               next: (request: HttpRequest<unknown>) => Observable<unknown>): Observable<unknown>
     {
-        this.ɵLogger = this.ɵLogger ?? this.injector.get(LOGGER, null);
+        const $this = this as unknown as ɵLogger;
+        $this.ɵLogger = $this.ɵLogger ?? this.injector.get(LOGGER, null);
 
-        if(!this.ɵLogger)
+        if(!$this.ɵLogger)
         {
             return next(request);
         }
 
-        this.ɵLogger.verbose(`RESTClient ${methodName}: Request {@request}`, 
+        $this.ɵLogger.verbose(`RESTClient ${methodName}: Request {@request}`, 
         {
             args,
             request
@@ -53,11 +55,11 @@ export class LoggerMiddleware implements RestMiddleware
         return next(request)
             .pipe(tap(response =>
                       {
-                          this.ɵLogger.verbose(`RESTClient ${methodName}: Response {@response}`, response);
+                          $this.ɵLogger?.verbose(`RESTClient ${methodName}: Response {@response}`, response);
                       },
                       error =>
                       {
-                          this.ɵLogger.verbose(`RESTClient ${methodName}: ErrorResponse {@error}`, error);
+                          $this.ɵLogger?.verbose(`RESTClient ${methodName}: ErrorResponse {@error}`, error);
                       }));
     }
 }
