@@ -1,11 +1,10 @@
-import {Type, Injector} from '@angular/core';
-import {HttpRequest, HttpResponse, HttpClient, HttpEvent} from '@angular/common/http';
-import {AdditionalInfo} from '@anglr/common';
+import {Type} from '@angular/core';
+import {HttpRequest, HttpResponse} from '@angular/common/http';
 import {StringDictionary, Dictionary} from '@jscrpt/common';
 import {Observable} from 'rxjs';
 
 import {ResponseType} from './responseType';
-import type {ParameterTransformFunc, ResponseTransformFunc} from '../rest/common';
+import type {ParameterTransformFunc, ResponseTransformFunc, RESTClient} from '../rest/common';
 
 /**
  * Type indicates that it should be removed from array
@@ -19,29 +18,6 @@ export class NotType<TType>
 }
 
 /**
- * Represents private defintion of rest client
- */
-export interface ɵRESTClient
-{
-    http: HttpClient;
-    baseUrl?: string;
-    injector?: Injector;
-    middlewaresOrder?: Type<RestMiddleware>[];
-    methodMiddlewares?: Type<RestMiddleware>[];
-    getBaseUrl(): string;
-    getDefaultHeaders(): string | {[name: string]: string | string[]};
-    requestInterceptor(req: HttpRequest<any>): HttpRequest<any>;
-    responseInterceptor<TBody = any>(res: Observable<HttpEvent<TBody>>): Observable<HttpEvent<any>>;
-}
-
-/**
- * Property descriptor that is used for creating decorators that can pass additional info to method
- */
-export interface AdditionalInfoPropertyDescriptor<TAdditional = any> extends TypedPropertyDescriptor<any>, AdditionalInfo<TAdditional>
-{
-}
-
-/**
  * Contains additional headers that will be added
  */
 export interface RestHttpHeaders extends TypedPropertyDescriptor<any>
@@ -49,7 +25,18 @@ export interface RestHttpHeaders extends TypedPropertyDescriptor<any>
     /**
      * Headers defintion to be added
      */
-    headers?: StringDictionary;
+    headers: StringDictionary;
+}
+
+/**
+ * Contains name of progress indicator group for local progress indicator
+ */
+export interface RestProgressIndicatorGroup extends TypedPropertyDescriptor<any>
+{
+    /**
+     * Name of progress indicator group
+     */
+    groupName: string;
 }
 
 /**
@@ -71,7 +58,7 @@ export interface RestResponseTransform extends TypedPropertyDescriptor<any>
     /**
      * Response transform function
      */
-    responseTransform?: ResponseTransformFunc;
+    responseTransform: ResponseTransformFunc;
 }
 
 /**
@@ -82,7 +69,7 @@ export interface RestDisabledInterceptors<TType = any> extends TypedPropertyDesc
     /**
      * Array of interceptors types that will be disabled
      */
-    disabledInterceptors?: Type<TType>[];
+    disabledInterceptors: Type<TType>[];
 }
 
 /**
@@ -97,17 +84,6 @@ export interface RestReportProgress extends TypedPropertyDescriptor<any>
 }
 
 /**
- * Contains indication whether is response full HttpResponse or just data
- */
-export interface RestFullHttpResponse extends TypedPropertyDescriptor<any>
-{
-    /**
-     * Indication whether is response full HttpResponse or just data
-     */
-    fullHttpResponse?: boolean;
-}
-
-/**
  * Contains data that are stored when REST method is set
  */
 export interface ɵRestMethod extends TypedPropertyDescriptor<any>
@@ -115,7 +91,7 @@ export interface ɵRestMethod extends TypedPropertyDescriptor<any>
     /**
      * Array of middlewares that are executed for each request
      */
-    middlewares?: RestMiddlewareRunMethod[];
+    middlewares: RestMiddlewareRunMethod[];
 }
 
 /**
@@ -126,7 +102,7 @@ export interface RestMethod extends TypedPropertyDescriptor<any>
     /**
      * Number of parameters that are on the method originaly
      */
-    originalParamsCount?: number;
+    originalParamsCount: number;
 }
 
 /**
@@ -149,14 +125,14 @@ export interface RestCaching extends TypedPropertyDescriptor<any>
      * Gets response from cache
      * @param request - Http request that is tested whether it is in cache
      */
-    getCachedResponse?: (request: HttpRequest<any>) => HttpResponse<any>|null;
+    getCachedResponse: (request: HttpRequest<any>) => HttpResponse<any>|null;
 
     /**
      * Saves response to cache for provided request and returns this response
      * @param request - Request that is identifies response
      * @param response - Response to be cached
      */
-    saveResponseToCache?: (request: HttpRequest<any>, response: HttpResponse<any>) => HttpResponse<any>;
+    saveResponseToCache: (request: HttpRequest<any>, response: HttpResponse<any>) => HttpResponse<any>;
 }
 
 /**
@@ -167,12 +143,12 @@ export interface KeyIndex
     /**
      * Key value that is passed to parameter
      */
-    key?: string;
+    key: string;
 
     /**
      * Index of parameter
      */
-    parameterIndex?: number;
+    parameterIndex: number;
 }
 
 /**
@@ -241,7 +217,7 @@ export interface RestMiddlewareRunMethod<TRequestBody = any, TResponseBody = any
      * @param request - Http request that you can modify
      * @param next - Used for calling next middleware with modified request
      */
-    (this: ɵRESTClient,
+    (this: RESTClient,
      id: string,
      target: TTarget,
      methodName: string,
@@ -252,6 +228,7 @@ export interface RestMiddlewareRunMethod<TRequestBody = any, TResponseBody = any
 
     /**
      * Runs code that is defined for this rest middleware, in this method you can modify request and response
+     * @param this - Method is bound to RESTClient
      * @param id - Unique id that identifies request method
      * @param target - Prototype of class that are decorators applied to
      * @param methodName - Name of method that is being modified
@@ -293,18 +270,11 @@ export interface BuildMiddlewaresFn
 {
     /**
      * Builds and returns array of middleware run functions
-     * @param middlewares - Array of set middleware types
-     * @param middlewaresOrder - Array of middleware types in order that should be executed
-     */
-    (middlewares: Type<RestMiddleware>[], middlewaresOrder: Type<RestMiddleware>[]): RestMiddlewareRunMethod[];
-
-    /**
-     * Builds and returns array of middleware run functions
      * @param this - Instance of RESTClient
      * @param middlewares - Array of set middleware types
      * @param middlewaresOrder - Array of middleware types in order that should be executed
      */
-    (this: ɵRestMethod, middlewares: Type<RestMiddleware>[], middlewaresOrder: Type<RestMiddleware>[]): RestMiddlewareRunMethod[]
+    (this: RESTClient, middlewares: Type<RestMiddleware>[], middlewaresOrder: Type<RestMiddleware>[]): RestMiddlewareRunMethod[]
 }
 
 /**
