@@ -1,5 +1,6 @@
-import {NgModule} from "@angular/core";
-import {WEB_SOCKET_HANDLE_PRODUCES_TYPE_PROVIDER, WEB_SOCKET_HANDLE_FILTER_NON_MATCHING_PROVIDER, WEB_SOCKET_HANDLE_RESPONSE_TRANSFORM_PROVIDER, WEB_SOCKET_HANDLE_RESULT_MIDDLEWARE_PROVIDER, WEB_SOCKET_HANDLE_STATUS_PROGRESS_INDICATOR_PROVIDER, WEB_SOCKET_HANDLE_STATUS_ERROR_COMPLETE_PROVIDER, WEB_SOCKET_HANDLE_STATUS_SUBSCRIBE_MIDDLEWARE_PROVIDER} from "./webSocketClient.extensibility";
+import {NgModule} from '@angular/core';
+
+import {WEB_SOCKET_HANDLE_PRODUCES_TYPE_PROVIDER, WEB_SOCKET_HANDLE_FILTER_NON_MATCHING_PROVIDER, WEB_SOCKET_HANDLE_RESPONSE_TRANSFORM_PROVIDER, WEB_SOCKET_HANDLE_RESULT_MIDDLEWARE_PROVIDER, WEB_SOCKET_HANDLE_STATUS_PROGRESS_INDICATOR_PROVIDER, WEB_SOCKET_HANDLE_STATUS_ERROR_COMPLETE_PROVIDER, WEB_SOCKET_HANDLE_STATUS_SUBSCRIBE_MIDDLEWARE_PROVIDER} from './webSocketClient.extensibility';
 
 /**
  * Module for WebSocket client providers, import only once, best at root level
