@@ -1,5 +1,6 @@
-import {InjectionToken} from "@angular/core";
-import {WebSocketHandleResultMiddleware, WebSocketHandleStatusSubscribeMiddleware} from "./webSocketClient.interface";
+import {InjectionToken} from '@angular/core';
+
+import {WebSocketHandleResultMiddleware, WebSocketHandleStatusSubscribeMiddleware} from './webSocketClient.interface';
 
 /**
  * Supported Produces response types
